DEA and simplification of HK76
The direct electrifying algorithm (DEA) is implemented. Also I made
some changes that simplify the Hoshen-Kopelman algorithm (and so the
DEA). I removed the vector contacts_cnt_point. It was used in the DEA
to make the elements. Instead I created a vector called elements.
Unlike the vector contacts_cnt_point, the vector elements does not need
to be sorted.