Update and new classes
I updated the classes for the Hoshen-Kopelman algorithm, for the direct
electrifying algorithm and for the Observation windows (Cutoff_Wins).

Also I created new classes:

Background grid: creates a kind of grid to determine the CNTs to trim
on each observation window. Also creates a vector called structure that
references all the 3D points

Contact grid: This is the grid with overlapping sub-regions that is
used to search for contacts

Percolation: This class checks for percolation on a single cluster