Done debugging App_Network_3D
The code seems to work fine, or at least it runs without errors.

In GenNetwork I commented the section where it exported tecplot files
so that I could run tests faster.

I added a new class, Printer, that can be used to save 1D and 2D
vectors to a text file. This can be used to debug the code or for to
save files that will be used as re-start files.

Modified the Makefile to include the new classes.

The rest of the modifications were corrections on the classes that
caused error or wrong outputs. There are also several commented
gout-statements that I used to debug the code.