Corrected error in generation algorithm
There was a small error in the generation algorithm when calculating
the length of a segment of a CNT inside the sample. Also if-statements
were modified to make the logic more clear.

In the comments, changed the word “original” to “initial” whenever it
was more suitable.

In the Clusters_fractions class I made a change when checking that the
length of CNTs in the clusters is the same as in the observation
window. Before it was just the difference in lengths and now it is the
difference normalized repeat to the length inside the observation
window. Finally, I increased the precision of the output when the
lengths are different. In some cases only the integer part was visible
so the error was not evident.