Found and corrected bug in Cutoff_wins
I found that some CNTs had only one point. I track down the bug to the
cutoff windows.

I introduced a new function to trim the CNTs. With this new function I
was able to get rid of that bug. In this update I keep both functions.
Right after this update I will make a second update where the old
function is deleted.

As a consequence of tracking the error, I introduced the following
changes:

- In Backbone_Network.cpp I added more information to the error messages
- In Direct_Electrifying.cpp there is a check for CNTs of size 1.
- In GenNetwork.cpp I modified the seeding for the random number
generator so that I could save into the output file the seed used. I
used only for debugging, so now this section is commented and the
previous way is the one in used. I will leave the second way commented
in case it is necessary to work with the same structure.

Other changes:

- In App_Network_3D.cpp I added a calculation for the time of each
iteration.
- In Backbone_Network.cpp I corrected a possible loss in precision when
comparing size_t with long int. Also, now the vector of currents is
saved to a file. In the header file I included some necessary header
files.
- In Clusters_fractions.cpp deleted unnecessary header file
- In Cutoff_Wins.cpp removed one unused argument from the function
Add_to_boundary_vectors