Correction in percolation, speed of of GenNetwork, additional capabilities of Printer
Corrected a function in Percolation.cpp that was supposed to consider
percolation by a single CNT. This could happen when the size of the
observation window in any direction is smaller than the CNT length.
Before, it was considering the size of the sample independently of the
size of the observation window. I made the correction so that it
considers the size of the observation window. Because of this, I also
had to make a small change in App_Network_3D.cpp.

In GenNetwork.cpp:
- I reduced the number of computations in a matrix-matrix
multiplication and a matrix-vector multiplication, by directly
calculating the elements of the resulting matrix and vector,
respectively.
- Added proper casting from integer to double when needed
- Added information to the output file when checking that a segment has
the correct orientation and when checking that a point is inside a
given cuboid
- I increased the maximum number of seeds (initial CNT points) from 1E6
to 1E9 in order to be able to generate larger samples (I could not
generate a sample of 40 microns with a limit of 1E6 seeds).

In Printer.cpp I increased the precision of the output value for
variables of type double. I also added a function to export CNT wires
for a given cluster of CNTs (i.e. a 1D vector that contains CNT numbers)

In Hoshen_Kopelman.cpp I just re-arranged the print functions and added
a few more. These calls to the printer functions are for debugging
purposes only, hence they are commented.