Fixed bug in random seeds
The problem that was causing the discrepancies in the code was due to
the initial point of the CNTs repeating. So several CNTs would
agglomerate around the same point.

To solve this, now there are three seeds for the initial point of a
CNT. One seed for every coordinate. Then every MAX_INT iterations the
srand() function is called to refresh the seeds. This because the
period of repetition was MAX_INT.