Changes in the boundary conditions for DEA
I modified the direct electrifying algorithm (DEA) so that there is a
different voltage on each boundary.

Also, I modified the zero-current cutoff to the one in the paper by Li
and Tchou. I also check if the gap is too large. In that case, I take
the criteria I was using before.