Zero cutoff update
I had forgotten to update this part since I had not uploaded this file
into Noor.

In this update I directly set the current cutoff as a drop in 9 orders
of magnitude in the current. I left the attempt I made on a combination
between the overall drop in magnitude of the current and the gap Li and
Chou found. However it did not worked.