One particle successfully generated
At this point and according to the visualizations, one particle is
generated correctly. At least this is inferred from the location of the
CNTs, the GNP is still missing from the visualizations.

The class GCH for the hybrid particle is new.

Changes:

In App_Network_3D:
- The new function that generates the network of hybrid particles was
updated.

In Input_Reader:
- A section for reading the GNP geometry was added
- In the RVE geometry section, there is the option to choose between
CNTs and hybrid particles
- In the nanotube geometry section, now the density of the CNTs is an
input

In GenNetwork:
- Most changes happened here
- The previous generation code was split into different function in
order to be able to generate a bunch of CNTs at a time. This bunch of
CNT corresponds to the top and bottom CNTs of a hybrid particle.