Hybrid particle and electrical resistance
The generation for the hybrid particle is finished. However it works
for small volume fraction (less than 1%).

Also the electrical resistance of the composite is calculated along
each direction. The class Electrical_analysis was added for this
purpose.

A few bugs were corrected that caused some resistances to be added
twice in the stiffness matrix for the DEA.