New generation for CNT, GNP, mixed and hybrid particles
The generation part of the code now can generate the following
particles:

1) Only CNTs
2) Only GNPs
3) Mixed CNTs and GNPS
4) CNT/GNP hybrid particles

The input file was modified so that the GNP geometry has more
parameters. Before this change some parameters were shared with the CNT
geometry.

However, the Hoshen-Kopelman and Direct electrifying algorithms only
work for CNTs and hybrid particles.

A bug was corrected that caused some resistances to be added
twice in the stiffness matrix for the DEA.